Change libhao to use size_t instead of HAO_INT.

+++ Lanczos/interface/include/Lanczos.h
@@ -5,9 +5,10 @@
 #ifndef AFQMCLIB_LANCZOS_H
 #define AFQMCLIB_LANCZOS_H
 
+#include <cstddef>
 class Lanczos
 {
-    int lanczosMatrixSize;
+    size_t lanczosMatrixSize;
  public:
 };
