+++ Lanczos/interface/test/LanczosBasisWfTest.cpp
@@ -7,7 +7,7 @@ using namespace tensor_hao;
 class LanczosBasisWfTest: public ::testing::Test
 {
  public:
-    const int L = 10;
+    const size_t L = 10;
     TensorHao<complex<double>, 1> vec;
     LanczosBasisWfTest( )
     {
@@ -77,9 +77,9 @@ TEST_F(LanczosBasisWfTest, normalize)
 
     TensorHao<complex<double>, 1> vecNorm(L);
     complex<double> inverseNorm(0,0);
-    for(int i = 0; i < L; ++i) inverseNorm += std::norm( vec(i) );
+    for(size_t i = 0; i < L; ++i) inverseNorm += std::norm( vec(i) );
     inverseNorm = 1.0 / sqrt(inverseNorm);
-    for(int i = 0; i < L; ++i) vecNorm(i) = vec(i) * inverseNorm;
+    for(size_t i = 0; i < L; ++i) vecNorm(i) = vec(i) * inverseNorm;
 
     EXPECT_FALSE( diff(vecNorm, wf.getWf(), 1e-12) );
 }
@@ -90,7 +90,7 @@ TEST_F(LanczosBasisWfTest, overlap)
     LanczosBasisWf wfLeft(vec), wfRight(vecRightSide);
     complex<double> overlap = wfLeft.calculateOverlapWith(wfRight);
     complex<double> overlap_exact;
-    for(int i = 0; i < L; ++i)
+    for(size_t i = 0; i < L; ++i)
     {
         overlap_exact += conj( vec(i) ) * vecRightSide(i);
     }
@@ -104,7 +104,7 @@ TEST_F(LanczosBasisWfTest, orthogonal)
 
     complex<double> overlap = wfBase.calculateOverlapWith(wf);
     TensorHao<complex<double>, 1> vecOrthogonal(L);
-    for(int i = 0; i < L; ++i)
+    for(size_t i = 0; i < L; ++i)
     {
         vecOrthogonal(i) = vec(i) - overlap * vecBase(i);
     }
